Update README.md
changed article months without capital letter to starting with capital letter, in my own language all months are written all lowercase, so I am not used to that. 

Removed the remark for "Article considered as fake" after the Slovenia article, because I mistakenly thought it was the Alipio article

There is still a 404 error  on:

 Covid-19, Cocooning, and Vitamin D Intake Requirements [Link](http://imj.ie/wp-content/uploads/2020/05/Covid-19-Cocooning-and-Vitamin-D-Intake-Requirements.pdf)

It wasn't clear to me what the dates are of the letters in the  "Comprehensive debate by the Irish Medical Journal" section, so I put date unknown there, you might want to check that.